Refactor Reconciler tests
The whole idea behind this change is to peek the actions emitted by the reconciliation process using the available API client, and validate
whether those actions are the expected for a given test scenario; in
other words, it validates if the right changes are being requested to
the API server, ensuring that the operator is following the expected
contract. As a result of that, most Reconciler tests are now sub-tests
within a single function.

* FakeDynClient() now returns a *fake.FakeDynamicClient type instead of
  dynamic.Interface. This gives access to the actions performed by the
  reconciliation run, increasing the reconciliation process'
  observability.

* Helper functions have been created to deal with common assertions,
  which simplified reasonably the test code.

* Most tests have been moved as cases within one test function.